requirements.txt
tensorflow==2.20.0
numpy==1.26.4
pillow==12.1.0
